# examples/03-petstore/spec/openapi.petstore.yaml
...
openapi: 3.2.0
...
info:
  title: Modern Petstore API
  summary: A modern, realistic pet store API demonstrating current best practices
  description: |-
    This is a modern Pet Store API based on the OpenAPI 3.2 specification. It demonstrates contemporary API design patterns and best practices including RESTful compliance, proper HTTP methods, structured error responses (RFC 9457), and modern authentication flows.

    Key features include pet catalog browsing, adoption workflow, order management, user accounts, AI-powered chat advisor, and real-time event notifications via webhooks. The API showcases OpenAPI 3.2 capabilities like hierarchical tags, QUERY HTTP method, OAuth device flow, and server-sent events.

    Some useful links:
    - [API Documentation](https://docs.petstoreapi.com)
    - [Source Code Repository](https://github.com/petstoreapi/PetstoreAPI)
    - [OpenAPI Specification](https://petstoreapi.com/v1/specs/modern-petstore-3.2.openapi.yaml)
  version: 1.0.0
  contact:
    name: Petstore API Support
    url: https://petstoreapi.com/support
    email: support@petstoreapi.com
  license:
    name: MIT
    identifier: MIT
...
paths:
  /pets/{id}:
    parameters:
      - name: id
        in: path
        description: Unique identifier for the pet
        required: true
        example: 01936c8f-1234-7000-8000-111111111111
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/TenantID'
    get:
      summary: Get Pet
      deprecated: false
      description: Retrieve detailed information about a specific pet.
      operationId: getPet
      tags:
        - Pet
      security: []
      x-codeSamples:
        - lang: TypeScript
          label: TypeScript SDK
          source: |-
            import { PetStoreAPI } from '@petstoreapi/sdk';

            const client = new PetStoreAPI({
              apiKey: process.env.PETSTORE_API_KEY
            });

            const pet = await client.pets.get('01936c8f-1234-7000-8000-111111111111');
            console.log(`Found ${pet.name}, a ${pet.age_months}-month-old ${pet.species}`);
        - lang: Python
          label: Python
          source: |-
            from petstore import PetStoreAPI

            client = PetStoreAPI(api_key=os.environ['PETSTORE_API_KEY'])

            pet = client.pets.get('01936c8f-1234-7000-8000-111111111111')
            print(f"Found {pet.name}, a {pet.age_months}-month-old {pet.species}")
        - lang: JavaScript
          label: JavaScript (Fetch)
          source: |-
            const response = await fetch(
              'https://api.petstoreapi.com/v1/pets/01936c8f-1234-7000-8000-111111111111',
              {
                headers: {
                  'Accept': 'application/json'
                }
              }
            );

            const pet = await response.json();
            console.log(`Found ${pet.name}, a ${pet.age_months}-month-old ${pet.species}`);
        - lang: Shell
          label: cURL
          source: |-
            curl -X GET 'https://api.petstoreapi.com/v1/pets/01936c8f-1234-7000-8000-111111111111' \
              -H 'Accept: application/json'
      responses:
        '200':
          description: Successful response with pet details
          headers:
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Remaining:
              $ref: '#/components/headers/RateLimit-Remaining'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
              examples:
                cat-example:
                  summary: A friendly cat
                  value:
                    id: 01936c8f-1234-7000-8000-111111111111
                    species: CAT
                    name: Whiskers
                    breed: Domestic Shorthair
                    ageMonths: 18
                    size: MEDIUM
                    color: Orange Tabby
                    gender: MALE
                    goodWithKids: true

                    price: '75.00'
                    description: Friendly and playful orange tabby looking for a loving home
                    status: AVAILABLE
                    photos:
                      - https://cdn.petstoreapi.com/pets/01936c8f-1234-7000-8000-111111111111/photo1.jpg
                      - https://cdn.petstoreapi.com/pets/01936c8f-1234-7000-8000-111111111111/photo2.jpg
                    medicalInfo:
                      spayedNeutered: true
                      vaccinated: true
                      microchipped: true
                      specialNeeds: false
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    put:
      summary: Update Pet
      deprecated: false
      description: Update information for an existing pet. Staff only.
      operationId: updatePet
      tags:
        - Pet
      security:
        - bearerAuth: []
        - oauth2:
            - write:pets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pet'
            example:
              name: Max Updated
              species: DOG
              breed: Golden Retriever
              ageMonths: 25
              size: LARGE
              color: Golden
              gender: MALE
              goodWithKids: true
              price: '275.00'
              currency: USD
              status: AVAILABLE
              description: Friendly and well-trained golden retriever
              medicalInfo:
                vaccinated: true
                spayedNeutered: true
                microchipped: true
                specialNeeds: false
                healthNotes: Up to date on all vaccinations, recent dental cleaning
      responses:
        '200':
          description: Pet updated successfully
          headers:
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Remaining:
              $ref: '#/components/headers/RateLimit-Remaining'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    delete:
      summary: Delete Pet
      deprecated: false
      description: Delete a pet from the system. Staff only.
      operationId: deletePet
      tags:
        - Pet
      security:
        - bearerAuth: []
        - oauth2:
            - write:pets
      responses:
        '204':
          description: Pet deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /pets:
    parameters:
      - $ref: '#/components/parameters/TenantID'
    post:
      summary: Create Pet
      deprecated: false
      description: >-
        Add a new pet to the store catalog, making it available for adoption.


        ## Pet Lifecycle Workflow


        When a new pet enters the system:


        1. **Intake**: Staff creates a pet record with this endpoint (status: `available`)

        2. **Profile**: Pet details include species, breed, age, medical info, photos, and adoption fee

        3. **Discovery**: Pet appears in search results and listings

        4. **Adoption Application**: Potential adopters can apply through the adoption endpoints

        5. **Adoption**: Once approved, pet status changes to `adopted`

        6. **Post-Adoption**: Pet record is retained for historical purposes


        **Access**: This operation requires staff permissions (`write:pets` scope or valid Bearer token). Only
        authenticated staff members can add pets to the system.
      operationId: createPet
      tags:
        - Pet
      security:
        - bearerAuth: []
        - oauth2:
            - write:pets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pet'
            example:
              name: Max
              species: DOG
              breed: Golden Retriever
              ageMonths: 24
              size: LARGE
              color: Golden
              gender: MALE
              goodWithKids: true
              price: '250.00'
              currency: USD
              status: AVAILABLE
              description: Friendly golden retriever looking for an active family
              medicalInfo:
                vaccinated: true
                spayedNeutered: true
                microchipped: true
                specialNeeds: false
                healthNotes: Up to date on all vaccinations
      x-codeSamples:
        - lang: TypeScript
          label: TypeScript SDK
          source: |-
            import { PetStoreAPI } from '@petstoreapi/sdk';

            const client = new PetStoreAPI({
              accessToken: process.env.OAUTH_ACCESS_TOKEN
            });

            const newPet = await client.pets.create({
              species: 'dog',
              name: 'Buddy',
              breed: 'Golden Retriever',
              ageMonths: 24,
              size: 'large',
              color: 'Golden',
              gender: 'male',
              goodWithKids: true,

              price: '150.00',
              description: 'Friendly golden retriever looking for an active family'
            });

            console.log(`Created pet with ID: ${newPet.id}`);
        - lang: Python
          label: Python
          source: |-
            from petstore import PetStoreAPI

            client = PetStoreAPI(access_token=os.environ['OAUTH_ACCESS_TOKEN'])

            new_pet = client.pets.create(
                species='dog',
                name='Buddy',
                breed='Golden Retriever',
                ageMonths=24,
                size='large',
                color='Golden',
                gender='male',
                goodWithKids=True,

                price='150.00',
                description='Friendly golden retriever looking for an active family'
            )

            print(f"Created pet with ID: {new_pet.id}")
        - lang: JavaScript
          label: JavaScript (Fetch)
          source: |-
            const response = await fetch(
              'https://api.petstoreapi.com/v1/pets',
              {
                method: 'POST',
                headers: {
                  'Authorization': `Bearer ${accessToken}`,
                  'Content-Type': 'application/json'
                },
                body: JSON.stringify({
                  species: 'dog',
                  name: 'Buddy',
                  breed: 'Golden Retriever',
                  ageMonths: 24,
                  size: 'large',
                  color: 'Golden',
                  gender: 'male',
                  goodWithKids: true,

                  price: '150.00',
                  description: 'Friendly golden retriever looking for an active family'
                })
              }
            );

            const newPet = await response.json();
            console.log(`Created pet with ID: ${newPet.id}`);
        - lang: Shell
          label: cURL
          source: |-
            curl -X POST 'https://api.petstoreapi.com/v1/pets' \
              -H 'Authorization: Bearer YOUR_ACCESS_TOKEN' \
              -H 'Content-Type: application/json' \
              -d '{
                "species": "dog",
                "name": "Buddy",
                "breed": "Golden Retriever",
                "ageMonths": 24,
                "size": "large",
                "color": "Golden",
                "gender": "male",
                "goodWithKids": true,

                "price": "150.00",
                "description": "Friendly golden retriever looking for an active family"
              }'
      responses:
        '201':
          description: Pet created successfully
          headers:
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Remaining:
              $ref: '#/components/headers/RateLimit-Remaining'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    get:
      summary: List Pets
      deprecated: false
      description: Retrieve a paginated list of pets available for adoption with optional filtering.
      operationId: listPets
      tags:
        - Pet
      security: []
      parameters:
        - name: species
          in: query
          description: Filter by pet species
          required: false
          schema:
            type: string
            enum:
              - DOG
              - CAT
              - RABBIT
              - BIRD
              - REPTILE
              - OTHER
        - name: status
          in: query
          description: Filter by adoption status
          required: false
          schema:
            type: string
            enum:
              - AVAILABLE
              - PENDING
              - ADOPTED
            default: AVAILABLE
        - name: ageMin
          in: query
          description: Minimum age in months
          required: false
          schema:
            type: integer
            minimum: 0
        - name: ageMax
          in: query
          description: Maximum age in months
          required: false
          schema:
            type: integer
            minimum: 0
        - name: size
          in: query
          description: Filter by pet size
          required: false
          schema:
            type: string
            enum:
              - SMALL
              - MEDIUM
              - LARGE
        - name: goodWithKids
          in: query
          description: Filter pets that are good with children
          required: false
          schema:
            type: boolean
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Successful response with pet collection
          headers:
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Remaining:
              $ref: '#/components/headers/RateLimit-Remaining'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetCollection'
              examples:
                multiple-pets:
                  summary: Multiple pets available
                  value:
                    data:
                      - id: 01936c8f-1234-7000-8000-111111111111
                        species: CAT
                        name: Whiskers
                        breed: Domestic Shorthair
                        ageMonths: 18
                        size: MEDIUM
                        color: Orange Tabby
                        gender: MALE
                        goodWithKids: true

                        price: '75.00'
                        description: Friendly and playful orange tabby looking for a loving home
                        status: AVAILABLE
                        photos:
                          - https://cdn.petstoreapi.com/pets/01936c8f-1234-7000-8000-111111111111/photo1.jpg
                      - id: 01936c8f-2345-7000-8000-222222222222
                        species: DOG
                        name: Max
                        breed: Labrador Retriever
                        ageMonths: 36
                        size: LARGE
                        color: Yellow
                        gender: MALE
                        goodWithKids: true

                        price: '150.00'
                        description: Energetic and loyal lab who loves fetch and long walks
                        status: AVAILABLE
                        photos:
                          - https://cdn.petstoreapi.com/pets/01936c8f-2345-7000-8000-222222222222/photo1.jpg
                    pagination:
                      page: 1
                      limit: 20
                      totalItems: 45
                      totalPages: 3
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    query:
      summary: Advanced Pet Search (QUERY method)
      deprecated: false
      description: >-
        Perform complex pet searches using the QUERY HTTP method (RFC draft-ietf-httpbis-safe-method-w-body). Unlike
        GET, this allows sending structured search criteria in the request body while maintaining safe, idempotent
        semantics.


        ## Why QUERY?


        The QUERY method is ideal for:

        - **Complex queries** that exceed URL length limits

        - **Structured search criteria** better expressed in JSON

        - **Safe operations** that don't modify resources

        - **Cacheable results** like GET requests


        ## Use Cases


        - Multi-criteria searches (medical history, temperament, location)

        - Saved search queries reused programmatically

        - Advanced filtering with nested conditions


        This endpoint returns the same response format as GET /pets but accepts complex search criteria in the body.
      operationId: searchPets
      tags:
        - Pet
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                criteria:
                  type: object
                  description: Search criteria for finding pets
                  properties:
                    species:
                      type: array
                      items:
                        type: string
                        enum:
                          - DOG
                          - CAT
                          - RABBIT
                          - BIRD
                          - REPTILE
                          - OTHER
                      description: Filter by one or more species
                    ageRange:
                      type: object
                      properties:
                        min:
                          type: integer
                          minimum: 0
                        max:
                          type: integer
                          minimum: 0
                    size:
                      type: array
                      items:
                        type: string
                        enum:
                          - SMALL
                          - MEDIUM
                          - LARGE
                    compatibility:
                      type: object
                      properties:
                        goodWithKids:
                          type: boolean

                    medical:
                      type: object
                      properties:
                        vaccinated:
                          type: boolean
                        spayedNeutered:
                          type: boolean
                        specialNeeds:
                          type: boolean
                sort:
                  type: object
                  properties:
                    field:
                      type: string
                      enum:
                        - ageMonths
                        - price
                    order:
                      type: string
                      enum:
                        - ASC
                        - DESC
                pagination:
                  type: object
                  properties:
                    page:
                      type: integer
                      minimum: 1
                      default: 1
                    limit:
                      type: integer
                      minimum: 1
                      maximum: 100
                      default: 20
            examples:
              family-friendly-dogs:
                summary: Family-friendly dogs
                value:
                  criteria:
                    species:
                      - DOG
                    ageRange:
                      min: 12
                      max: 72
                    size:
                      - medium
                      - large
                    compatibility:
                      goodWithKids: true

                    medical:
                      vaccinated: true
                      spayedNeutered: true
                  sort:
                    field: age_months
                    order: asc
                  pagination:
                    page: 1
                    limit: 20
      responses:
        '200':
          description: Search results
          headers:
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Remaining:
              $ref: '#/components/headers/RateLimit-Remaining'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Pet'
                  pagination:
                    $ref: '#/components/schemas/PaginationInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /pets/{id}/images:
    post:
      summary: Upload Pet Image
      deprecated: false
      description: Upload an image for the specified pet. Requires authentication and write:pets permission.
      operationId: uploadPetPhoto
      tags:
        - Pet
      parameters:
        - name: id
          in: path
          description: Unique identifier for the pet
          required: true
          example: 01936c8f-1234-7000-8000-111111111111
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                additionalMetadata:
                  description: Additional data to pass to server
                  type: string
                  example: ''
                file:
                  description: File to upload
                  type: string
                  format: binary
                  example: ''
            examples: {}
      responses:
        '201':
          description: Operation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 3
                type: ex dolor
                message: Pecus carcer cometes credo.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - bearerAuth: []
        - oauth2:
            - write:pets
...
components:
  schemas:
    Pet:
      type: object
      description: Animal information in the pet store available for adoption.
      required:
        - id
        - name
        - species
        - ageMonths
        - price
        - currency
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
          description: Unique identifier for the pet (UUID v7)
          examples:
            - 019b4132-70aa-764f-b315-e2803d882a24
            - 019b4127-54d5-76d9-b626-0d4c7bfce5b6
        species:
          type: string
          enum:
            - DOG
            - CAT
            - RABBIT
            - BIRD
            - REPTILE
            - OTHER
          description: The species of the pet
        name:
          type: string
          minLength: 1
          maxLength: 50
          description: The pet's name
          examples:
            - Whiskers
            - Max
            - Luna
        breed:
          type: string
          description: The breed of the pet
          examples:
            - Domestic Shorthair
            - Labrador Retriever
            - Holland Lop
        ageMonths:
          type: integer
          minimum: 0
          description: Age of the pet in months
          examples:
            - 18
            - 36
            - 6
        size:
          type: string
          enum:
            - SMALL
            - MEDIUM
            - LARGE
          description: Size category of the pet
        color:
          type: string
          description: Primary color or coloring pattern
          examples:
            - Orange Tabby
            - Black
            - Brown and White
        gender:
          type: string
          enum:
            - MALE
            - FEMALE
            - UNKNOWN
          description: The pet's gender
        goodWithKids:
          type: boolean
          description: Whether the pet is good with children

        price:
          type: string
          # format: double removed as it is not valid for string
          description: Adoption fee amount
          examples:
            - '75.00'
            - '150.00'
            - '50.00'
        currency:
          type: string
          description: Currency code for the adoption fee (ISO 4217)
          default: USD
          pattern: ^[A-Z]{3}$
          examples:
            - USD
            - EUR
            - GBP
        description:
          type: string
          description: Detailed description of the pet's personality and traits
        status:
          type: string
          enum:
            - AVAILABLE
            - PENDING
            - ADOPTED
            - NOT_AVAILABLE
          description: Current adoption status
        createdAt:
          type: string
          format: date-time
          readOnly: true
          description: Timestamp when the pet record was created (RFC 3339)
          examples:
            - '2025-12-21T13:56:23Z'
            - '2025-11-15T08:30:00Z'
        updatedAt:
          type: string
          format: date-time
          readOnly: true
          description: Timestamp when the pet record was last updated (RFC 3339)
          examples:
            - '2025-12-21T13:56:23Z'
            - '2025-12-21T15:30:45Z'
        tenantId:
          type: string
          format: uuid
          readOnly: true
          description: >-
            Optional tenant identifier for data isolation. When present, indicates this pet belongs to a specific
            tenant. Null or omitted means the pet is in the shared/public data pool.
          examples:
            - 550e8400-e29b-41d4-a716-446655440000
            - 7c9e6679-7425-40de-944b-e07fc1f90ae7
          nullable: true
        photos:
          type: array
          items:
            type: string
            format: uri
          description: URLs of pet photos
          minItems: 0
        medicalInfo:
          type: object
          properties:
            spayedNeutered:
              type: boolean
              description: Whether the pet has been spayed or neutered
            vaccinated:
              type: boolean
              description: Whether the pet is up to date on vaccinations
            microchipped:
              type: boolean
              description: Whether the pet has a microchip
            specialNeeds:
              type: boolean
              description: Whether the pet has special medical needs
            healthNotes:
              type: string
              description: Additional medical notes
          unevaluatedProperties: false
      unevaluatedProperties: false
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            pet-not-found:
              summary: Pet not found
              value:
                type: https://petstoreapi.com/errors/not-found
                title: Not Found
                status: 404
                detail: The requested pet with ID '123' was not found.
                instance: /pets/123
...
  securitySchemes:
    oauth2:
      type: oauth2
      description: OAuth 2.0 authorization with multiple flows including device authorization for IoT devices and smart TVs
      flows:
        authorizationCode:
          authorizationUrl: https://auth.petstoreapi.com/authorize
          tokenUrl: https://auth.petstoreapi.com/token
          refreshUrl: https://auth.petstoreapi.com/refresh
          scopes:
            read:pets: Read pet information
            write:pets: Create and update pets
            read:orders: Read order information
            write:orders: Create and manage orders
            read:user: Read user profile
            write:user: Update user profile
            chat:write: Access AI chat completions
        deviceCode:
          deviceAuthorizationUrl: https://auth.petstoreapi.com/device/authorize
          tokenUrl: https://auth.petstoreapi.com/token
          refreshUrl: https://auth.petstoreapi.com/refresh
          scopes:
            read:pets: Read pet information
            chat:write: Access AI chat completions
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Bearer token authentication using JWT (JSON Web Token). Include the token in the Authorization header as:
        `Authorization: Bearer <token>`
  headers:
    RateLimit-Limit:
      description: The maximum number of requests allowed in the current time window
      schema:
        type: integer
      example: 100
    RateLimit-Remaining:
      description: The number of requests remaining in the current time window
      schema:
        type: integer
      example: 95
    RateLimit-Reset:
      description: The time at which the current rate limit window resets (Unix timestamp)
      schema:
        type: integer
      example: 1702648800
  parameters:
    TenantID:
      name: X-Tenant-ID
      in: header
      description: >-
        Optional tenant identifier for data isolation. When provided, all operations will be scoped to this tenant,
        ensuring data separation between different organizations or users. If omitted, operations will access the
        shared/public data pool where data may be visible to and modified by other users.
      required: false
      schema:
        type: string
        format: uuid
      examples:
        organization:
          value: 550e8400-e29b-41d4-a716-446655440000
          summary: Organization tenant ID
        development:
          value: 7c9e6679-7425-40de-944b-e07fc1f90ae7
          summary: Development environment
        production:
          value: 9b5e4f89-3c12-4a5e-9f8e-2d3c5a7e6b1a
          summary: Production environment
...
servers:
  - url: https://api.petstoreapi.com/v1
    description: Production server
security:
  - bearerAuth: []
  - oauth2: []
...